test/cxl-poison.sh: test inject and clear poison by region offset
The CXL kernel driver recently added support to inject and clear
poison in a region by specifying an offset. Add a test case to the
existing cxl-poison unit test that demonstrates how to use the new
debugfs attributes. Use the kernel trace log to validate the round
trip address translations.

SKIP, do not fail, if the new debugfs attributes are not present.

See the kernel ABI documentation for usage:
Documentation/ABI/testing/debugfs-cxl

[as: rm whitespace in a comment]

Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/20250823025954.3161194-1-alison.schofield@intel.com
Signed-off-by: Alison Schofield <alison.schofield@intel.com>